feat: add serde primitive, update command
