+++ src/encoding/mod.rs
@@ -2,6 +2,7 @@ mod errors;
 #[cfg(feature = "evm")]
 pub mod evm;
 pub mod models;
+pub mod serde_primitives;
 pub mod strategy_encoder;
 mod swap_encoder;
 pub mod tycho_encoder;
